Update myStyle.scss
Made a lot of progress on the app development. Added a left sidebar that opens when the left "Browse" icon is pressed. Changed the color theme. Optimized and added a few more animations. The design at this point is almost final and optimized for the final stage (as I imagine it right now), but may suffer changes until the end of the project,