Update myScript.js
Added and implemented a lot of methods and interactions with the interface. Added a left browsing sidebar where the "Photo Albums" will be accessible in the future. Added new animations that are triggered by interactions with the "Browse" and "Realign" buttons (icons). Next important step is to add and complete the 3 photo albums arrays. Got to find a secure way to host the pictures that will be linked in the albums.