Update myScript.js
Added new methods for the "browse albums" sidebar, that sets the chosen album as the "pictures" state of the main component. Now on picture-open the sidebar browse nav is closed(collapsed). Added a new "Intro" component for the "intro" animation. The next step is to add all the links for the photos in the 3 Photo Albums arrays and that will be the final important part of the project.